Seperate AstraCAN.h into .h/.cpp
